split gsheet javascript and add hackathon description
