+++ index.html
@@ -583,8 +583,8 @@
   </script>
   <script type="module" src="https://unpkg.com/ionicons@4.5.10-0/dist/ionicons/ionicons.esm.js"></script>
   <script nomodule="" src="https://unpkg.com/ionicons@4.5.10-0/dist/ionicons/ionicons.js"></script>
-  <script type="module" src="assets/js/main.js">
-  </script>
+  <script type="module" src="assets/js/main.js"></script>
+  <script type="module" src="assets/js/gsheet.js"></script>
 
 </body>
