+++ roadshow-rewards.html
@@ -91,7 +91,7 @@
             <br>
 
             <div class="row intro-text">
-                <p>This year's IndabaX South Africa will be held as a Roadshow of various
+                <p>This year's Indaba𝕏 South Africa will be held as a Roadshow of various
                     fun events, <br>
                     taking place from <b>3 September to 27 November</b>
                 </p>
@@ -102,7 +102,7 @@
                 </div>
                 <div class="col-md-4 col-xs-12">
                     <h3>
-                        <a class="btn btn-warning btn-block" href="http://eepurl.com/hCE_C5">
+                        <a class="btn btn-warning btn-block" href="http://eepurl.com/hCE_C5" target="_blank">
                             Sign up to the roadshow
                         </a>
                     </h3>
@@ -119,12 +119,25 @@
         <div class="container">
             <div class="row">
                 <div class="col-md-8 col-xs-12">
-                    <h2>Hackathon</h2>
+                    <h2>Hackathon Hero(es)</h2>
                     <h3>R20 000 prize pool</h3>
-                    <p>About the hackathon</p>
+                    <p>Do you have what it takes to predict whether or not a customer is about to take an item from a
+                        shelf?
+                        The Cape AI team is pleased to announce that they are partnering with Zindi on their latest
+                        hackathon for IndabaX.
+                        Test your ability to read shoppers and be rewarded with a prize from our R20 000 prize pot.
+                        The Cape AI x Indaba𝕏 Hackathon is all about predicting whether or not a stream of 30 video
+                        frames from a real customer’s shopping session contains a “take” action.
+                    </p>
+                    <p><em>Participation in the hackathon is restricted to Indaba𝕏 attendees and individuals currently
+                            living in South Africa. Full rules on the Zindi competition page.</em></p>
                     <a class="btn btn-success btn-lg" href="#TODO">
-                        Sign up to the hackathon
+                        Sign up to the <strong>hackathon</strong> with your code*
                     </a>
+
+                    <p><small>*the code is available once you've signed up to our <a class="text-white"
+                                href="http://eepurl.com/hCE_C5" target="_blank">mailing
+                                list</a> - sent via email and posted on the community forum</small></p>
                 </div>
             </div>
             <!-- row -->
@@ -217,6 +230,19 @@
                     </p>
                 </div>
             </div>
+            <div class="row">
+                <div class="col-md-12 cols-xs-12">
+                    <p class="site-info"><a
+                            href="https://docs.google.com/document/d/1Xt8JwUM7GIJUKii8O2Z4-pplaZAQZHDipjGxzdp9xkY/edit?usp=sharing"
+                            target="_blank">Privacy policy</a>
+                        |
+                        <a href="https://docs.google.com/document/d/1mzOwxhOufeVGq5uhtE_ZtbObVqjbWbKwabSyO-7ql5g/edit?usp=sharing"
+                            target="_blank">Code
+                            of
+                            conduct</a>
+                    </p>
+                </div>
+            </div>
         </div>
     </footer>
 
@@ -229,6 +255,7 @@
     </script>
     <script type="module" src="https://unpkg.com/ionicons@4.5.10-0/dist/ionicons/ionicons.esm.js"></script>
     <script nomodule="" src="https://unpkg.com/ionicons@4.5.10-0/dist/ionicons/ionicons.js"></script>
+    <script type="module" src="assets/js/main.js"></script>
 
 </body>
